requirements.txt
pymssql
pandas
openpyxl
azure-storage-blob
openai
pymupdf
